+++ src/backend/src/Application/AI/VectorStore/VectorQuery.cs
@@ -2,7 +2,7 @@ namespace backend.Application.AI.VectorStore;
 
 public class VectorQuery
 {
-    public float[] Vector { get; set; } = null!;
+    public double[] Vector { get; set; } = null!;
     public int TopK { get; set; }
     public Dictionary<string, string>? Filter { get; set; }
 }
