+++ src/backend/src/Application/Faces/Commands/FaceDetectionResultDto.cs
@@ -8,5 +8,5 @@ public class FaceDetectionResultDto
     public BoundingBoxDto BoundingBox { get; set; } = null!;
     public float Confidence { get; set; }
     public string? Thumbnail { get; set; } // Base64 encoded image
-    public float[]? Embedding { get; set; } // Changed from List<float>
+    public double[]? Embedding { get; set; }
 }
