+++ src/backend/src/Domain/Entities/TextChunk.cs
@@ -5,7 +5,7 @@ namespace backend.Domain.Entities
         public string Id { get; set; } = Guid.NewGuid().ToString();
         public string Content { get; set; } = null!;
         public Dictionary<string, string> Metadata { get; set; } = new Dictionary<string, string>();
-        public float[]? Embedding { get; set; }
+        public double[]? Embedding { get; set; }
         public float Score { get; set; }
     }
 }
