+++ src/backend/src/Infrastructure/AI/Embeddings/OpenAIEmbeddingProvider.cs
@@ -1,9 +1,9 @@
+using System.Net.Http.Headers;
+using System.Text.Json;
 using backend.Application.Common.Interfaces;
 using backend.Application.Common.Models;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
-using System.Net.Http.Headers;
-using System.Text.Json;
 
 namespace backend.Infrastructure.AI.Embeddings;
 
@@ -23,11 +23,11 @@ public class OpenAIEmbeddingProvider : IEmbeddingProvider
         _logger = logger;
     }
 
-    public async Task<Result<float[]>> GenerateEmbeddingAsync(string text, CancellationToken cancellationToken = default)
+    public async Task<Result<double[]>> GenerateEmbeddingAsync(string text, CancellationToken cancellationToken = default)
     {
         if (string.IsNullOrWhiteSpace(_settings.OpenAI.ApiKey))
         {
-            return Result<float[]>.Failure("OpenAI API key is not configured.");
+            return Result<double[]>.Failure("OpenAI API key is not configured.");
         }
 
         if (text.Length > MaxTextLength)
@@ -57,24 +57,24 @@ public class OpenAIEmbeddingProvider : IEmbeddingProvider
                 if (root.TryGetProperty("data", out JsonElement dataElement) && dataElement.GetArrayLength() > 0)
                 {
                     JsonElement embeddingElement = dataElement[0].GetProperty("embedding");
-                    float[] embedding = embeddingElement.EnumerateArray().Select(e => (float)e.GetDouble()).ToArray();
-                    return Result<float[]>.Success(embedding);
+                    double[] embedding = embeddingElement.EnumerateArray().Select(e => e.GetDouble()).ToArray();
+                    return Result<double[]>.Success(embedding);
                 }
             }
 
-            return Result<float[]>.Failure("Failed to parse embedding from OpenAI API response.");
+            return Result<double[]>.Failure("Failed to parse embedding from OpenAI API response.");
         }
         catch (HttpRequestException ex)
         {
-            return Result<float[]>.Failure($"OpenAI API request failed: {ex.Message}");
+            return Result<double[]>.Failure($"OpenAI API request failed: {ex.Message}");
         }
         catch (JsonException ex)
         {
-            return Result<float[]>.Failure($"Failed to deserialize OpenAI API response: {ex.Message}");
+            return Result<double[]>.Failure($"Failed to deserialize OpenAI API response: {ex.Message}");
         }
         catch (Exception ex)
         {
-            return Result<float[]>.Failure($"An unexpected error occurred during OpenAI embedding generation: {ex.Message}");
+            return Result<double[]>.Failure($"An unexpected error occurred during OpenAI embedding generation: {ex.Message}");
         }
     }
 }
